requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
jinja2==3.1.5
python-multipart==0.0.20
sqlalchemy==2.0.36
pydantic==2.10.5
pydantic-settings==2.7.1
python-dotenv==1.0.1
apscheduler==3.10.4
yfinance==0.2.54
pandas==2.2.3
requests-cache==1.2.1
